Fix build crash while using MSFT to build OVMF, also fix some build warning report for PCD type issues.

git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@12202 6f19259b-4bc3-4df7-8a09-765794883524

+++ OvmfPkg/PlatformPei/Xen.c
@@ -66,8 +66,8 @@ XenConnect (
   AsmCpuid (XenLeaf + 1, &XenVersion, NULL, NULL, NULL);
   DEBUG ((EFI_D_ERROR, "Detected Xen version %d.%d\n",
           XenVersion >> 16, XenVersion & 0xFFFF));
-  mXenInfo.VersionMajor = XenVersion >> 16;
-  mXenInfo.VersionMinor = XenVersion & 0xFFFF;
+  mXenInfo.VersionMajor = (UINT16)(XenVersion >> 16);
+  mXenInfo.VersionMinor = (UINT16)(XenVersion & 0xFFFF);
 
   /* TBD: Locate hvm_info and reserve it away. */
   mXenInfo.HvmInfo = NULL;
